backend/requirements.txt
Flask==3.1.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
requests==2.31.0
urllib3==2.0.7
openai==0.28.1

